requirements.txt
APScheduler==3.5.1
dateparser==0.7.0
Flask==0.12.2
Flask-And-Redis==0.7
Flask-APScheduler==1.8.0
Flask-Celery-Helper==1.1.0
Flask-Cors==3.0.3
Flask-JWT==0.3.2
Flask-JWT-Extended==3.3.4
Flask-Mail==0.9.1
Flask-PyMongo==0.5.1
Flask-Redis==0.3.0
Flask-SocketIO==3.0.0
gunicorn==19.7.1
PyJWT==1.4.2
pymongo==3.5.1
pytz==2017.3
redis==2.10.6
regex==2018.2.21
requests==2.18.4
urllib3==1.22
webargs==1.8.1
Werkzeug==0.12.2
raven